third_party/blink/web_tests/external/wpt/css/css-values/calc-size/calc-size-parsing-expected.txt

This is a testharness.js-based test.
Found 3 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['max-width'] = "calc-size(max-content, size / 2)" should set the property value
  assert_equals: serialization should be canonical expected "calc-size(max-content, 0.5 * size)" but got "calc-size(max-content, size * 0.5)"
[FAIL] e.style['max-height'] = "calc-size(fit-content, 30px + size / 2)" should set the property value
  assert_equals: serialization should be canonical expected "calc-size(max-content, 30px + 0.5 * size)" but got "calc-size(fit-content, 30px + size * 0.5)"
[FAIL] e.style['width'] = "calc-size(fit-content, 50% + size / 2)" should set the property value
  assert_equals: serialization should be canonical expected "calc-size(max-content, 50% + 0.5 * size)" but got "calc-size(fit-content, 50% + size * 0.5)"
Harness: the test ran to completion.
